+++ flutter_inappwebview_forge_platform_interface/lib/src/in_app_browser/platform_in_app_browser.dart
@@ -1861,6 +1861,13 @@ In that case, after the `window.addEventListener("flutterInAppWebViewPlatformRea
         apiUrl:
             'https://developer.android.com/reference/android/webkit/WebChromeClient#onGeolocationPermissionsShowPrompt(java.lang.String,%20android.webkit.GeolocationPermissions.Callback)',
       ),
+      IOSPlatform(
+        apiName:
+            'WKUIDelegate.webView(_:requestGeolocationPermissionFor:initiatedByFrame:decisionHandler:)',
+        apiUrl:
+            'https://developer.apple.com/documentation/webkit/wkuidelegate/webview(_:requestgeolocationpermissionfor:initiatedbyframe:decisionhandler:)',
+        available: '26.0',
+      ),
     ],
   )
   FutureOr<GeolocationPermissionShowPromptResponse?>?