+++ flutter_inappwebview_forge_platform_interface/lib/src/in_app_browser/platform_in_app_browser.g.dart
@@ -1302,6 +1302,7 @@ enum PlatformInAppBrowserEventsMethod {
   ///
   ///**Officially Supported Platforms/Implementations**:
   ///- Android WebView ([Official API - WebChromeClient.onGeolocationPermissionsShowPrompt](https://developer.android.com/reference/android/webkit/WebChromeClient#onGeolocationPermissionsShowPrompt(java.lang.String,%20android.webkit.GeolocationPermissions.Callback)))
+  ///- iOS WKWebView 26.0+ ([Official API - WKUIDelegate.webView(_:requestGeolocationPermissionFor:initiatedByFrame:decisionHandler:)](https://developer.apple.com/documentation/webkit/wkuidelegate/webview(_:requestgeolocationpermissionfor:initiatedbyframe:decisionhandler:)))
   ///
   ///**Parameters - Officially Supported Platforms/Implementations**:
   ///- [origin]: all platforms
@@ -2399,6 +2400,7 @@ extension _PlatformInAppBrowserEventsMethodSupported
         return ((kIsWeb && platform != null) || !kIsWeb) &&
             [
               TargetPlatform.android,
+              TargetPlatform.iOS,
             ].contains(platform ?? defaultTargetPlatform);
       case PlatformInAppBrowserEventsMethod.onJsAlert:
         return ((kIsWeb && platform != null) || !kIsWeb) &&